fixed all of the incorrect references to "rootArtifactId" in the controllers
For this sdk-212 jira ticket we have found very simple errors in the code specifically in the __moduleClassnamePrefix__Controller files. We have changed incorrect references to "rootArtifactId" such as "rootartifactId" and "rootArtifactid". We then tested the sdk by creating both reference applications and platform modules and observed their controllers which are now being generated correctly.